Spell optimization stage 7, and the review defects it was waiting on
Stage 7 -- Spell is recycled through a per-thread freelist.

It is fixed-size and built once per cast, once per auto-shot and once per
triggered chain link: the most frequently constructed non-trivial object in the
server. Class-level operator new/delete keep up to 64 blocks per thread; the
virtual destructor is what makes the queue delete through a BasicEvent* land in
Spell operator delete rather than the global one.

Last, deliberately. A pool turns a lifetime bug from a crash into silent
corruption -- the block is still readable, still the right size, and now belongs
to a different cast. It is only safe because stage 1 made ownership a unique_ptr
and because the raw pointers the pipeline kept across a cast are gone. Which is
what the rest of this commit is.

The heartbeat holder dereferenced a caster the same constructor treats as
nullable fifty lines earlier, so a fear or a polymorph landing after its
original caster left the map took the world down.

Spell::finish iterated the live m_modAuras list while casting spells into it.
RemoveAura calls list::remove, which frees the node, and the next ++i walked it.
Snapshot first and pin each holder, the same shape the proc system in Unit.cpp
already uses; the loop also stops early on an aura its own trigger removed.

Every SpellModifier a player had active leaked at logout. AURA_REMOVE_BY_DELETE
skips the unapply, so the only delete in the tree never ran, ~Aura was empty and
~Player never touches m_spellMods -- and the list kept a pointer into freed aura
memory on the way out. ~Aura releases it now, silently, because by then there
may be no session to send a modifier update to. HandleAddModifier also nulls the
member after AddSpellMod frees it, which is what let ~Aura tell "already
released" from "never released", and closes a double free on any second unapply.

m_selfContainer was written by SetCurrentCastedSpell and read nowhere. Upstream
read it in ~Spell to null a current-spell slot pointing at a spell being
destroyed. Restored: the flag discipline should make it unreachable, which is
precisely what a defence like this is for.

The effect dispatcher narrowed a uint32 DBC field to uint8 before bound-checking
it, so 256 + n would have dispatched as n.

+++ src/game/WorldHandlers/SpellEffectDispatch.cpp
@@ -93,7 +93,11 @@ void Spell::HandleEffects(Unit* pUnitTarget, Item* pItemTarget, GameObject* pGOT
     itemTarget = pItemTarget;
     gameObjTarget = pGOTarget;
 
-    uint8 eff = m_spellInfo->Effect[i];
+    // Effect is a uint32 in the DBC. Narrowing it before the bound check meant a
+    // value of 256 + n passed the check as n and dispatched the wrong handler
+    // through SpellEffects[]. 1.12 data never reaches 256, which was the only
+    // thing making that safe.
+    uint32 eff = m_spellInfo->Effect[i];
 
     damage = int32(CalculateDamage(i, unitTarget) * DamageMultiplier);
